Don't re-download files that already exist
If the download script fails i.e. `requests.exceptions.ConnectionError: (MaxRetryError("HTTPSConnectionPool` it starts all over again and re-downloads all of the files again. This update checks to see if the file exists (from a previous run) and will skip over any existing files and only download files that need to be downloaded.